pyproject.toml
[project]
name = "market-sentiment-oracle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "alpaca-py>=0.43.2",
    "altair<5",
    "backtrader>=1.9.78.123",
    "google-cloud-storage>=3.8.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas<3",
    "pandera>=0.28.1",
    "plotly>=6.5.2",
    "pyarrow>=23.0.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "streamlit>=1.30.0",
    "tensorflow>=2.20.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.0.0",
    "urllib3>=1.26.18,<2.0.0",
    "yfinance>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
